UFF-041 - criação de testes e refatoração do castle para BoardCastleExtensions
